Add Permissions management

+++ backend/src/resolvers/Mutation.js
@@ -2,14 +2,24 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
+const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
-    //TODO Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
 
     const item = await ctx.db.mutation.createItem(
       {
         data: {
+          //this is how to creaste a relationshop between item and user
+          user: {
+            connect: {
+              id: ctx.request.userId
+            }
+          },
           ...args
         }
       },
@@ -34,9 +44,15 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     //1. find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user {id}}`);
     //2. check if they own that item, or have the permission
-    //TODO:
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if (!(ownsItem && hasPermission))
+      throw new Error('You are not allowed to delete this item!!!');
+
     //4. delete item
     return ctx.db.mutation.deleteItem(
       {
@@ -98,7 +114,7 @@ const Mutations = {
     if (!user) {
       throw new Error(`No such uyser found for email ${args.email}`);
     }
-    //2. set a reset token and expirey on the user
+    //2. set a reset token and expiry on the user
     const randomBytesPromisefied = promisify(randomBytes);
     const resetToken = (await randomBytesPromisefied(20)).toString('hex');
     const resetTokenExpiry = Date.now() + 3600000; // 1 hour from now
@@ -107,10 +123,20 @@ const Mutations = {
       data: { resetToken, resetTokenExpiry }
     });
     //3. email them that reset token
-    console.log(res);
+    await transport.sendMail({
+      from: 'test@kyoui.com',
+      to: user.email,
+      subject: 'Your password reset token',
+      html: makeANiceEmail(
+        `Your password token is here!! \n\n <a href="${
+          process.env.FRONTEND_URL
+        }/reset?resetToken=${resetToken}">Click Here to Reset</a>`
+      )
+    });
+    //4. return the message
     return { message: 'Thanks' };
   },
-  async resetPassword(parent, args, ctx, info) {
+  async reset(parent, args, ctx, info) {
     //1. check if password match
     if (args.password !== args.confirmPassword) {
       throw new Error("Yo Password don't match!");
@@ -147,6 +173,38 @@ const Mutations = {
     //8. return the new user
     return updatedUser;
     //9
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    //1. check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to udpate permissions');
+    }
+    //2. query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+    //3. check if they have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONSUPDATE']);
+    //4. update ther permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
